+++ docs/source/news.rst
@@ -1,6 +1,13 @@
 What's New
 ==========
 
+0.10.0
+^^^^^^
+
+Features:
+
+* ``extract_fields_for_exceptions`` allows for more useful :ref:`logging of failed actions<extract errors>` by extracting additional fields from exceptions.
+
 0.9.0
 ^^^^^